@@ -1,12 +1,13 @@
 apply plugin: 'com.android.library'
+
 apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
@@ -26,8 +27,9 @@ android {
 dependencies {
     api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
-    api 'com.android.support:design:27.1.0'
-    api 'com.android.support:recyclerview-v7:27.1.0'
+    api "androidx.recyclerview:recyclerview:1.0.0"
+
+    api 'com.google.android.material:material:1.0.0'
 }
 
 apply from: 'https://raw.githubusercontent.com/Commit451/gradle-android-javadocs/1.0.0/gradle-android-javadocs.gradle'
