@@ -2,21 +2,21 @@ package com.commit451.modalbottomsheetdialogfragment
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.support.annotation.LayoutRes
-import android.support.annotation.MenuRes
-import android.support.design.widget.BottomSheetDialogFragment
-import android.support.v4.app.FragmentManager
-import android.support.v7.view.menu.MenuBuilder
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.MenuInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.annotation.LayoutRes
+import androidx.annotation.MenuRes
+import androidx.appcompat.view.menu.MenuBuilder
+import androidx.fragment.app.FragmentManager
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.commit451.modalbottomsheetdialogfragment.ModalBottomSheetDialogFragment.Builder
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 @SuppressLint("RestrictedApi")
 /**
@@ -64,7 +64,7 @@ class ModalBottomSheetDialogFragment : BottomSheetDialogFragment() {
         val arguments = arguments
                 ?: throw IllegalStateException("You need to create this via the builder")
 
-        val optionHolders = arguments.getParcelableArrayList<OptionHolder>(KEY_OPTIONS)
+        val optionHolders = arguments.getParcelableArrayList<OptionHolder>(KEY_OPTIONS)!!
 
         val options = mutableListOf<Option>()
 
@@ -79,10 +79,10 @@ class ModalBottomSheetDialogFragment : BottomSheetDialogFragment() {
             }
         }
 
-        adapter = Adapter({
+        adapter = Adapter {
             listener?.onModalOptionSelected(this@ModalBottomSheetDialogFragment.tag, it)
             dismissAllowingStateLoss()
-        })
+        }
         adapter.layoutRes = arguments.getInt(KEY_LAYOUT)
         adapter.header = arguments.getString(KEY_HEADER)
         adapter.headerLayoutRes = arguments.getInt(KEY_HEADER_LAYOUT_RES)
